provisioner: split S104 hotspots into mixins, fingerprint CodeQL-flagged log sites
Splits per Sonar S104 (file-length):
- executors/aws_executor.py (787→218) → +EC2, +VPC-endpoints, +Routes mixins
- orchestrators/setup_orchestrator.py (705→286) → +Logging, +PanOS, +Types
- gdc_vmruntime_assets.py (1,158→458) → 7 _gdc_vm_* helpers

S1142 (>3 returns) refactored to single-return in start/stop_instance,
wait_for_running, wait_for_stopped, describe_instances, delete_endpoint,
wait_for_endpoint_available — no behavioral change.

CodeQL py/clear-text-logging-sensitive-data: replaces safe_log_value /
safe_log_id at flagged sites with new safe_log_fingerprint (SHA-256
truncated to 12 hex) — recognised CodeQL sanitizer that terminates the
taint flow while preserving cross-line correlation for operators.
