+++ shifter/engine/provisioner/cloud/aws/secrets.py
@@ -9,7 +9,7 @@ from botocore.exceptions import BotoCoreError, ClientError
 
 from cloud.aws.base import BaseAWSAdapter
 from cloud.exceptions import CloudSecretsError
-from log_redact import safe_log_id
+from log_redact import safe_log_fingerprint
 
 logger = logging.getLogger(__name__)
 
@@ -20,7 +20,7 @@ class AWSSecretsStore(BaseAWSAdapter):
     _service_name = "secretsmanager"
 
     def get_secret(self, secret_id: str) -> str:
-        logger.debug("get_secret: secret_id=%s", safe_log_id(secret_id))
+        logger.debug("get_secret: secret_id_fp=%s", safe_log_fingerprint(secret_id))
         try:
             client = self._get_client()
             response = client.get_secret_value(SecretId=secret_id)
@@ -29,5 +29,5 @@ class AWSSecretsStore(BaseAWSAdapter):
             # Binary secret — base64-decode to string
             return base64.b64decode(response["SecretBinary"]).decode("utf-8")
         except (ClientError, BotoCoreError) as e:
-            logger.exception("get_secret: failed secret_id=%s", safe_log_id(secret_id))
+            logger.exception("get_secret: failed secret_id_fp=%s", safe_log_fingerprint(secret_id))
             raise CloudSecretsError(f"Failed to retrieve secret: {e}") from e
