+++ shifter/engine/provisioner/cloud/gcp/secrets.py
@@ -6,7 +6,7 @@ import logging
 
 from cloud.exceptions import CloudSecretsError
 from cloud.gcp.base import build_secret_version_name, import_google_module
-from log_redact import safe_log_id
+from log_redact import safe_log_fingerprint
 
 logger = logging.getLogger(__name__)
 
@@ -16,7 +16,7 @@ class GCPSecretsStore:
 
     @staticmethod
     def get_secret(secret_id: str) -> str:
-        logger.debug("get_secret: secret_id=%s", safe_log_id(secret_id))
+        logger.debug("get_secret: secret_id_fp=%s", safe_log_fingerprint(secret_id))
         try:
             secretmanager = import_google_module("google.cloud.secretmanager")
             client = secretmanager.SecretManagerServiceClient()
@@ -25,5 +25,5 @@ class GCPSecretsStore:
         except ImportError as e:
             raise CloudSecretsError("GCP secrets support requires google-cloud-secret-manager") from e
         except Exception as e:
-            logger.exception("get_secret: failed secret_id=%s", safe_log_id(secret_id))
+            logger.exception("get_secret: failed secret_id_fp=%s", safe_log_fingerprint(secret_id))
             raise CloudSecretsError(f"Failed to retrieve GCP secret: {e}") from e
