+++ sonar-project.properties
@@ -72,36 +72,13 @@ sonar.sourceEncoding=UTF-8
 
 # Rule suppressions
 # python:S1722 ("Add inheritance from object") conflicts with ruff UP004 which
-# *removes* explicit object inheritance as redundant in Python 3. The repo
-# treats ruff as the authoritative style; suppress S1722 globally so the two
-# rules don't fight on every commit.
-#
-# python:S1142 ("too many returns") on executors/aws_executor.py: every public
-# operation here wraps a boto3 call in try/except WaiterError/except ClientError/
-# except Exception and returns a CommandResult per branch. The 3-return budget
-# would require either collapsing exception handlers into a dispatch helper
-# (loses the per-exception logging fidelity SOC has come to rely on) or
-# introducing a result-variable single-return pattern that buys 1 return
-# but adds 5 lines of plumbing per function. Suppressed at the file level
-# until a broader executor refactor consolidates the pattern.
-sonar.issue.ignore.multicriteria=s1722,s1142_aws_executor,s104_aws_executor,s104_setup_orchestrator,s104_provisioner_db,s104_gdc_vmruntime
+# *removes* explicit object inheritance as redundant in Python 3. Both rules
+# cannot be satisfied simultaneously — one says add ``(object)``, the other
+# strips it on every save. The repo treats ruff as the authoritative style
+# (it auto-formats on pre-commit), so S1722 is suppressed globally.
+sonar.issue.ignore.multicriteria=s1722
 sonar.issue.ignore.multicriteria.s1722.ruleKey=python:S1722
 sonar.issue.ignore.multicriteria.s1722.resourceKey=**/*.py
-sonar.issue.ignore.multicriteria.s1142_aws_executor.ruleKey=python:S1142
-sonar.issue.ignore.multicriteria.s1142_aws_executor.resourceKey=**/executors/aws_executor.py
-# python:S104 ("file too long") on four provisioner files PR #838 touched only
-# to add log-redaction wrappers. The lines actually added by this PR are small;
-# the underlying files are pre-existing tech debt slated for a dedicated split
-# tracked in a follow-up issue. Suppress here so the gate stays honest while
-# the split lands separately.
-sonar.issue.ignore.multicriteria.s104_aws_executor.ruleKey=python:S104
-sonar.issue.ignore.multicriteria.s104_aws_executor.resourceKey=**/executors/aws_executor.py
-sonar.issue.ignore.multicriteria.s104_setup_orchestrator.ruleKey=python:S104
-sonar.issue.ignore.multicriteria.s104_setup_orchestrator.resourceKey=**/orchestrators/setup_orchestrator.py
-sonar.issue.ignore.multicriteria.s104_provisioner_db.ruleKey=python:S104
-sonar.issue.ignore.multicriteria.s104_provisioner_db.resourceKey=**/engine/provisioner/provisioner_db.py
-sonar.issue.ignore.multicriteria.s104_gdc_vmruntime.ruleKey=python:S104
-sonar.issue.ignore.multicriteria.s104_gdc_vmruntime.resourceKey=**/engine/provisioner/gdc_vmruntime_assets.py
 
 # File-header enforcement (ADR-015). Web:HeaderCheck compares the start of every
 # HTML file against this literal block. Django `{# #}` comment syntax is used so
